requirements.txt
absl-py==0.7.0
asn1crypto==0.24.0
astor==0.7.1
attrs==17.4.0
Automat==0.6.0
backcall==0.1.0
bleach==3.1.0
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloud-init==18.3
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
curlify==2.1.1
decorator==4.3.2
defusedxml==0.5.0
distro-info==0.18
Django==2.1.7
entrypoints==0.3
facebook-business==3.2.8
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
hibagent==1.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
Markdown==3.0.1
MarkupSafe==1.0
mistune==0.8.4
mock==2.0.0
mysql-connector==2.1.6
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.2
oauthlib==2.0.6
PAM==0.4.2
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.0
ptyprocess==0.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
Pygments==2.3.1
pygobject==3.26.1
PyJWT==1.5.3
PyMySQL==0.9.3
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.2
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2018.9
pyxdg==0.25
PyYAML==3.12
pyzmq==18.0.0
qtconsole==4.4.3
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==0.20.3
scipy==1.2.1
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==16.0.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.18
ssh-import-id==5.7
systemd-python==234
tensorboard==1.13.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==6.0.1
traitlets==4.3.2
Twisted==17.9.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
zope.interface==4.3.2
